refactor: trying to refactor migration
