+++ lib/parser.js
@@ -21,7 +21,7 @@ export const parseIndex = (idx) => {
 
   if (idx.method) options.indexType = idx.type; // Set a type for the index, e.g. BTREE. See the documentation of the used dialect
 
-  if (idx.parser && idx.parser !== '') options.parser = idx.parser; // For FULLTEXT columns set your parser
+  if (idx.parser) options.parser = idx.parser; // For FULLTEXT columns set your parser
 
   idx.options = options;
 
@@ -81,7 +81,6 @@ export const getColumnTypeNameBase = (col, prefix = 'Sequelize.') => {
       ret = [ret];
 
       if (options.zerofill) ret.push('ZEROFILL');
-
       if (options.unsigned) ret.push('UNSIGNED');
 
       return `${ret.join('.')}`;
